#8 init fnc save time match card, change fnc findAllBySectionId, refr…

+++ kuro/src/main/java/com/hus/englishapp/kuro/service/ResultTestService.java
@@ -1,10 +1,12 @@
 package com.hus.englishapp.kuro.service;
 
+import com.hus.englishapp.kuro.exception.AppException;
 import com.hus.englishapp.kuro.model.Ques;
 import com.hus.englishapp.kuro.model.ResultTest;
 import com.hus.englishapp.kuro.model.dto.ResultTestResponseDto;
 import com.hus.englishapp.kuro.repository.ResultTestRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -15,23 +17,36 @@ import java.util.List;
 public class ResultTestService {
     @Autowired
     private ResultTestRepository resultTestRepository;
+    @Autowired
+    private UserService userService;
 
     @Transactional
-    public List<ResultTest> create(String sectionId, String userId, List<String> quesIds) {
-        resultTestRepository.deleteBySectionIdAndUserId(sectionId, userId);
-        List<ResultTest> listResultWrongQues = new ArrayList<>();
-        for (String quesId: quesIds) {
-            ResultTest resultTest = new ResultTest();
-            resultTest.setSectionId(sectionId);
-            resultTest.setUserId(userId);
-            resultTest.setWrongQuesId(quesId);
-            resultTestRepository.save(resultTest);
-            listResultWrongQues.add(resultTest);
+    public List<ResultTest> create(String sectionId, List<String> quesIds) {
+        try {
+            String userId = userService.currUser().getId();
+            if (userId == null) {
+                throw new AppException("Không tìm thấy tài khoản người dùng hiện tại", HttpStatus.UNAUTHORIZED);
+            }
+
+            resultTestRepository.deleteBySectionIdAndUserId(sectionId, userId);
+            List<ResultTest> listResultWrongQues = new ArrayList<>();
+
+            for (String quesId : quesIds) {
+                ResultTest resultTest = new ResultTest();
+                resultTest.setSectionId(sectionId);
+                resultTest.setUserId(userId);
+                resultTest.setWrongQuesId(quesId);
+                resultTestRepository.save(resultTest);
+                listResultWrongQues.add(resultTest);
+            }
+            return listResultWrongQues;
+        } catch (Exception e) {
+            throw new RuntimeException(e.getMessage());
         }
-        return listResultWrongQues;
     }
 
-    public ResultTestResponseDto findAllBySectionIdAndUserId(String sectionId, String userId) {
+    public ResultTestResponseDto findAllBySectionIdAndUserId(String sectionId) {
+        String userId = userService.currUser().getId();
         return resultTestRepository.findAllToGetQues(sectionId, userId);
     }
 }
